+++ forms/temp/forms.html
@@ -7,9 +7,9 @@
    </style>
     </head>
     <body>
-        <form action="/upload_text" method="post" enctype="multipart/form-data">
+        <form action="/upload_text/" method="post" enctype="multipart/form-data">
          <h3> Insert text: <br> </h3>
-        <input type="text" name="name"  id="user" size= "100" > <br>
+        <input type="text" name="name"  id="user" size="100" maxlength="5000"> <br>
         <button type="sumbit"> load text  <br> </button>
         <button type="reset"> reset text  <br> </button>
         </form>
